feat: Call the firePostTransformActions method on all transformers and refactor diagostics messages into another file
