+++ src/Scripts/build.ts
@@ -3,7 +3,7 @@ import { randomUUID } from 'crypto';
 import * as fs from 'fs';
 import { Builder, parseStringPromise } from 'xml2js';
 import { TSUtilities } from '../Utilities/TSUtilities';
-import ts from 'typescript';
+import { firePostTransformActions } from '../Utilities/TransformerUtilities';
 
 const [path, bin, command, ...args] = process.argv;
 
@@ -95,53 +95,7 @@ export async function build(): Promise<string[]> {
             ]
         });
 
-        // Get and store all diagnostic messages generated by typescript to display them when the task
-        // finishes
-        const allDiagnostics = ts.getPreEmitDiagnostics(program).concat(emitResult.diagnostics);
-        const defaultFormatHost: ts.FormatDiagnosticsHost = {
-            getCurrentDirectory: () => ts.sys.getCurrentDirectory(),
-            getCanonicalFileName: fileName => fileName,
-            getNewLine: () => ts.sys.newLine
-        };
-
-        const formattedDiagnostics = ts.formatDiagnosticsWithColorAndContext(allDiagnostics, defaultFormatHost);
-
-        // If an error causes a compilation failure, display the diagnostic messages and fail
-        if (emitResult.emitSkipped) {
-            process.stdout.write(`\r\x1b[1;31m✖\x1b[0m Failed building ${projectName || 'project'}\n`);
-
-            console.log(formattedDiagnostics);
-
-            throw new Error('Compilation failed.');
-        }
-
-        // Validate thingworx-specific constraints
-        for (const key in twConfig.store) {
-            if (key.startsWith('@')) continue;
-
-            const transformer = twConfig.store[key];
-            transformer.validateConstraints();
-        }
-
-        const diagnosticMessages = twConfig.store['@diagnosticMessages'] as unknown as DiagnosticMessage[];
-
-        // If any errors were reported, display them and fail
-        if (diagnosticMessages.some(m => m.kind == DiagnosticMessageKind.Error)) {
-            process.stdout.write(`\r\x1b[1;31m✖\x1b[0m Failed building ${projectName || 'project'}\n`);
-
-            for (const message of diagnosticMessages) {
-                switch (message.kind) {
-                    case DiagnosticMessageKind.Error:
-                        console.log(`🛑 \x1b[1;31mError\x1b[0m ${message.message}`);
-                        break;
-                    case DiagnosticMessageKind.Warning:
-                        console.log(`🔶 \x1b[1;33mWarning\x1b[0m ${message.message}`);
-                        break;
-                }
-            }
-
-            throw new Error('Validation failed.');
-        }
+        const { tsDiagnostics, validationDiagnostics } = firePostTransformActions(program, twConfig.store, projectName, emitResult);
 
         // Store an array of all transformers created, to be used to extract
         // debug information for the whole project
@@ -210,15 +164,15 @@ export async function build(): Promise<string[]> {
 
         fs.writeFileSync(`${outPath}/metadata.xml`, outXML);
 
-        process.stdout.write(`\r\x1b[1;32m✔\x1b[0m Built ${projectName || 'project'}${formattedDiagnostics.length ? ' (with warnings):' : '     '}\n`);
+        process.stdout.write(`\r\x1b[1;32m✔\x1b[0m Built ${projectName || 'project'}${validationDiagnostics.length ? ' (with warnings):' : '     '}\n`);
 
         // Write out the diagnostic messages at the end of the task
-        if (formattedDiagnostics) {
-            console.log(formattedDiagnostics);
+        if (tsDiagnostics) {
+            console.log(tsDiagnostics);
         }
 
         // If any validation errors were reported, display them at the end of the task
-        for (const message of diagnosticMessages) {
+        for (const message of validationDiagnostics) {
             console.log(`🔶 \x1b[1;33mWarning\x1b[0m ${message.message}`);
         }
     }
